Run linters and plan for the 0.9.0 release.
Update the MSRV to 1.63.0 and fix old lints, add default class implementations, and other clippy enhancements.

+++ lexical-write-float/src/compact.rs
@@ -374,7 +374,7 @@ pub unsafe fn generate_digits(
             idx += 1;
         }
 
-        part1 -= digit as u64 * div;
+        part1 -= digit * div;
         kappa -= 1;
 
         let tmp = (part1 << -one.exp) + part2;
